requirements.txt
openai==1.13.3
pypdf==4.0.2
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.31.1
python-docx==1.1.0